Move AlpacaToMessages to other transforms
